+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/controller/admin/PrisonExcludeDatesController.kt
@@ -17,7 +17,9 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import uk.gov.justice.digital.hmpps.visitscheduler.config.ErrorResponse
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.ExcludeDateDto
+import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.SessionScheduleWithDateExclusionsDto
 import uk.gov.justice.digital.hmpps.visitscheduler.service.PrisonConfigService
+import uk.gov.justice.digital.hmpps.visitscheduler.service.SessionTemplateService
 import java.time.LocalDate
 
 const val PRISONS_PATH: String = "/prisons"
@@ -26,12 +28,15 @@ const val ADD_PRISON_EXCLUDE_DATE: String = "$PRISON_EXCLUDE_DATE_PATH/add"
 const val REMOVE_PRISON_EXCLUDE_DATE: String = "$PRISON_EXCLUDE_DATE_PATH/remove"
 const val GET_PRISON_EXCLUDE_DATES: String = PRISON_EXCLUDE_DATE_PATH
 
+const val GET_SESSION_EXCLUDE_DATES: String = "$PRISONS_PATH/{prisonCode}/config/session-templates/exclude-dates/future"
+
 @RestController
 @Validated
 @Tag(name = "Prison exclude dates rest controller")
 @RequestMapping(name = "Prison Exclude Dates Configuration Resource", produces = [MediaType.APPLICATION_JSON_VALUE])
 class PrisonExcludeDatesController(
   private val prisonConfigService: PrisonConfigService,
+  private val sessionTemplateService: SessionTemplateService,
 ) {
   @PreAuthorize("hasRole('VISIT_SCHEDULER')")
   @PutMapping(ADD_PRISON_EXCLUDE_DATE)
@@ -140,4 +145,32 @@ class PrisonExcludeDatesController(
     @PathVariable
     prisonCode: String,
   ): List<ExcludeDateDto> = prisonConfigService.getExcludeDates(prisonCode)
+
+  @PreAuthorize("hasRole('VISIT_SCHEDULER')")
+  @GetMapping(GET_SESSION_EXCLUDE_DATES)
+  @Operation(
+    summary = "Get all current or future exclude dates by session for a prison.",
+    description = "Get all current or future exclude dates by session for a prison. Returns only sessions that are blocked in the future.",
+    responses = [
+      ApiResponse(
+        responseCode = "200",
+        description = "All sessions that are blocked - current or future are returned, returns an empty list if no sessions are blocked in the future (including today's sessions)",
+      ),
+      ApiResponse(
+        responseCode = "401",
+        description = "Unauthorized to access this endpoint",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
+      ),
+      ApiResponse(
+        responseCode = "403",
+        description = "Incorrect permissions to get sessions that are blocked in the future",
+        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
+      ),
+    ],
+  )
+  fun getFutureExcludedSessionsForPrison(
+    @Schema(description = "prison code", example = "BHI", required = true)
+    @PathVariable
+    prisonCode: String,
+  ): List<SessionScheduleWithDateExclusionsDto> = sessionTemplateService.getFutureExcludedSessionsForPrison(prisonCode)
 }
