+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/controller/admin/SessionTemplateExcludeDatesController.kt
@@ -17,16 +17,12 @@ import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import uk.gov.justice.digital.hmpps.visitscheduler.config.ErrorResponse
 import uk.gov.justice.digital.hmpps.visitscheduler.dto.ExcludeDateDto
-import uk.gov.justice.digital.hmpps.visitscheduler.dto.sessions.SessionScheduleWithDateExclusionsDto
 import uk.gov.justice.digital.hmpps.visitscheduler.service.SessionTemplateService
 import java.time.LocalDate
 
 private const val SESSION_TEMPLATES_PATH: String = "/session-templates"
-const val SESSION_TEMPLATE_EXCLUDE_DATE_PATH: String = "$SESSION_TEMPLATE_PATH/{reference}/exclude-date"
-const val SESSION_TEMPLATE_FUTURE_EXCLUDE_DATES_FOR_PRISON_PATH: String = "$SESSION_TEMPLATES_PATH/{prisonCode}/exclude-dates/future"
-
+const val SESSION_TEMPLATE_EXCLUDE_DATE_PATH: String = "$SESSION_TEMPLATES_PATH/{reference}/exclude-date"
 const val ADD_SESSION_TEMPLATE_EXCLUDE_DATE: String = "$SESSION_TEMPLATE_EXCLUDE_DATE_PATH/add"
-
 const val REMOVE_SESSION_TEMPLATE_EXCLUDE_DATE: String = "$SESSION_TEMPLATE_EXCLUDE_DATE_PATH/remove"
 const val GET_SESSION_TEMPLATE_EXCLUDE_DATES: String = SESSION_TEMPLATE_EXCLUDE_DATE_PATH
 
@@ -141,32 +137,4 @@ class SessionTemplateExcludeDatesController(
     @PathVariable
     reference: String,
   ): List<ExcludeDateDto> = sessionTemplateService.getExcludeDates(reference)
-
-  @PreAuthorize("hasRole('VISIT_SCHEDULER')")
-  @GetMapping(SESSION_TEMPLATE_FUTURE_EXCLUDE_DATES_FOR_PRISON_PATH)
-  @Operation(
-    summary = "Get all current or future exclude dates by session for a prison.",
-    description = "Get all current or future exclude dates by session for a prison. Returns only sessions that are blocked in the future.",
-    responses = [
-      ApiResponse(
-        responseCode = "200",
-        description = "All sessions that are blocked - current or future are returned, returns an empty list if no sessions are blocked in the future (including today's sessions)",
-      ),
-      ApiResponse(
-        responseCode = "401",
-        description = "Unauthorized to access this endpoint",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "403",
-        description = "Incorrect permissions to get sessions that are blocked in the future",
-        content = [Content(mediaType = "application/json", schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  fun getFutureExcludedSessionsForPrison(
-    @Schema(description = "prison code", example = "BHI", required = true)
-    @PathVariable
-    prisonCode: String,
-  ): List<SessionScheduleWithDateExclusionsDto> = sessionTemplateService.getFutureExcludedSessionsForPrison(prisonCode)
 }
